rabbitMq + scheduled is working

+++ src/main/java/ru/itmo/businesslogic/services/QuestionService.java
@@ -3,6 +3,7 @@ package ru.itmo.businesslogic.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.itmo.businesslogic.dao.QuestionDao;
+import ru.itmo.businesslogic.dao.UserDao;
 import ru.itmo.businesslogic.dto.QuestionDto;
 import ru.itmo.businesslogic.entities.Question;
 import ru.itmo.businesslogic.enums.Tag;
@@ -12,6 +13,9 @@ public class QuestionService {
 
     @Autowired
     private QuestionDao questionDao;
+
+    @Autowired
+    private UserDao userDao;
 
     public QuestionDto addQuestion(QuestionDto questionDto){
 
@@ -46,10 +50,23 @@ public class QuestionService {
         return questionDao.getAllQuestions();
     }
 
-    public QuestionDto changeQuestionStatus(Integer questionId, boolean isValid) {
+    public QuestionDto getQuestionsForEvaluation(String token) {
+        return questionDao.getQuestionsForEvaluation(token);
+    }
+
+
+
+    public QuestionDto changeQuestionStatus(Integer questionId, boolean isValid, String token) {
         if(questionId==null)
         {
             return new QuestionDto("Bad request no id found");
+        }
+        Question question = questionDao.findQuestionById(questionId);
+        if(question==null){
+            return new QuestionDto("No question with id = "+questionId);
+        }
+        if(!userDao.findUserById(question.getModeratorId()).getToken().equals(token)) {
+            return new QuestionDto("You can't evaluate this question");
         }
         if(!isValid) {
             return questionDao.deleteQuestion(questionId);
